refactor: Modularize configuration files and update imports for improved organization
