+++ src/config/index.ts
@@ -1,10 +1,10 @@
 /**
- * Configuration type definitions
- * Re-exports from defaults for convenience
+ * Configuration module
+ * Re-exports all configuration from modular files
  */
 
+// Defaults
 export type { CaptionAlignment } from "./defaults.ts";
-
 export {
     DEFAULT_VIDEO_DIMENSIONS,
     DEFAULT_VIDEO_SETTINGS,
@@ -15,11 +15,13 @@ export {
     DEFAULT_PATHS,
 } from './defaults.ts';
 
+// Provider types and configs
+export type { AIProvider, AIImageProvider, ProviderConfig } from './providers.ts';
+export { PROVIDER_CONFIGS } from './providers.ts';
+
+// Service configs
 export {
-    type AIProvider,
-    type AIImageProvider,
-    type ProviderConfig,
-    PROVIDER_CONFIGS,
+    parseIdList,
     TELEGRAM,
     TRANSCRIPTION,
     AI_TEXT,
@@ -27,7 +29,7 @@ export {
     MINIO,
     PATHS,
     DEBUG,
-    validateEnvironment,
-    getAIConfig,
-    parseIdList,
-} from "./environment.ts";
+} from './services.ts';
+
+// Validation
+export { validateEnvironment, getAIConfig } from './validation.ts';
