+++ src/services/llm/generator.ts
@@ -6,7 +6,7 @@
  * Phase 3: Verify and auto-correct queries for consistency
  */
 
-import { AI_TEXT, getAIConfig } from '../../config/environment.ts';
+import { AI_TEXT, getAIConfig } from '../../config/index.ts';
 import type { ImageSearchQuery } from '../../types/index.ts';
 import type { ResolvedStyle } from '../../styles/types.ts';
 import * as logger from '../../utils/logger.ts';
